pyproject.toml
[tool.poetry]
name = "python-telegram-bot-seed"
version = "0.1.0"
description = "Skeleton project for implementing bots in Python using the module python-telegram-bot"
authors = ["Ale Sanchez Medina <alesanchezmedina@gmail.com>"]
license = "The Unlicense"

[tool.poetry.dependencies]
python = "^3.8"
python-telegram-bot = "^20.3"
python-dotenv = "^1.0.0"
sqlalchemy = "^2.0.15"
alembic = "^1.11.1"

[tool.poetry.dev-dependencies]
pylint = "^2.12.2"
black = "^22.1.0"
isort = "^5.10.1"
bandit = "^1.7.4"

[tool.poetry.group.dev.dependencies]
autopep8 = "^2.0.2"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
